requirements.txt
# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# LangChain 核心
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# MCP 对接
langchain-mcp-adapters>=0.1.0

# 向量缓存
pgvector>=0.3.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Nacos 服务发现
nacos-sdk-python>=0.1.8

# 工具库
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
httpx>=0.27.0
openai>=1.50.0
